Lab 3

+++ backend/src/game/game.controller.ts
@@ -9,8 +9,10 @@ import {
   UseInterceptors,
 } from '@nestjs/common';
 import { GameService } from './game.service';
+import { Authentication, CognitoUser } from '@nestjs-cognito/auth';
 
 @Controller('game')
+@Authentication()
 export class GameController {
   constructor(private readonly gameService: GameService) {}
 
@@ -21,43 +23,62 @@ export class GameController {
   }
 
   @Post('create')
-  createGame(@Body() body: { name: string; userName: string }) {
+  createGame(@CognitoUser() user: any, @Body() body: { name: string }) {
     if (!body.name) {
       throw new HttpException('Name is required', 400);
     }
-    const game = this.gameService.createGame(body.name, body.userName);
+    if (!user.nickname) {
+      throw new HttpException('User not found', 404);
+    }
+    const game = this.gameService.createGame(
+      body.name,
+      user.nickname,
+      user.sub,
+    );
     return {
       id: game.id,
-      playerPrivateKey: game.player1,
-      playerPublicKey: game.player1pub,
     };
   }
 
   @Post('quit-any')
-  quitAnyGame(@Body() body: { playerPrivateKey: string }) {
-    this.gameService.removePlayerFromAllGames(body.playerPrivateKey);
+  quitAnyGame(@CognitoUser() user: any) {
+    if (!user.sub) {
+      throw new HttpException('User not found', 404);
+    }
+    this.gameService.removePlayerFromAllGames(user.sub);
     this.gameService.removeEmptyGames();
     return { message: 'Game quit' };
   }
 
+  @Get('player-id')
+  getPlayerId(@CognitoUser() user: any) {
+    if (!user.sub) {
+      throw new HttpException('User not found', 404);
+    }
+    return {
+      playerId: user.sub,
+    };
+  }
+
   @Post('join/:id')
-  joinGame(@Param('id') id: string, @Body() body: { userName: string }) {
+  joinGame(@Param('id') id: string, @CognitoUser() user: any) {
     const game = this.gameService.findOne(id);
     if (!game) {
       throw new HttpException('Game not found', 404);
     }
+    if (!user.sub) {
+      throw new HttpException('User not found', 404);
+    }
+    if (game.player1 === user.sub) {
+      throw new HttpException('You are already in the game', 400);
+    }
     if (this.gameService.isGameAvailable(game.id)) {
-      const playerData = this.gameService.addPlayerToGame(
-        game.id,
-        body.userName,
-      );
+      this.gameService.addPlayerToGame(game.id, user.nickname, user.sub);
       if (this.gameService.isGameReady(game)) {
         this.gameService.startGame(game);
       }
       return {
         id: game.id,
-        playerPrivateKey: playerData.privateKey,
-        playerPublicKey: playerData.pubKey,
       };
     }
     throw new HttpException('Game not available', 400);
@@ -73,10 +94,9 @@ export class GameController {
   @Post(':id/move')
   makeMove(
     @Param('id') id: string,
+    @CognitoUser() user: any,
     @Body()
     body: {
-      playerPrivateKey: string;
-      playerPublicKey: string;
       row: number;
       column: number;
     },
@@ -88,27 +108,13 @@ export class GameController {
     if (game.status !== 'in-progress') {
       throw new HttpException('Game not in progress', 400);
     }
-    if (game.currentPlayer !== body.playerPublicKey) {
+    if (!user.sub) {
+      throw new HttpException('User not found', 404);
+    }
+    if (game.currentPlayer !== user.sub) {
       throw new HttpException('Not your turn', 400);
     }
-    if (
-      !this.gameService.authPlayer(
-        game.id,
-        body.playerPrivateKey,
-        body.playerPublicKey,
-      )
-    ) {
-      throw new HttpException('Auth fail', 400);
-    }
-    if (
-      this.gameService.makeMove(
-        game,
-        body.row,
-        body.column,
-        body.playerPrivateKey,
-        body.playerPublicKey,
-      )
-    ) {
+    if (this.gameService.makeMove(game, body.row, body.column, user.sub)) {
       return this.gameService.findOne(id);
     }
     throw new HttpException('Invalid move', 400);
@@ -116,46 +122,34 @@ export class GameController {
 
   @Post(':id/restart')
   @UseInterceptors(ClassSerializerInterceptor)
-  restartGame(
-    @Param('id') id: string,
-    @Body() body: { playerPrivateKey: string; playerPublicKey: string },
-  ) {
+  restartGame(@Param('id') id: string, @CognitoUser() user: any) {
     const game = this.gameService.findOne(id);
     if (!game) {
       throw new HttpException('Game not found', 404);
     }
-    if (
-      !this.gameService.authPlayer(
-        game.id,
-        body.playerPrivateKey,
-        body.playerPublicKey,
-      )
-    ) {
-      throw new HttpException('Auth fail', 400);
+    if (!user.sub) {
+      throw new HttpException('User not found', 404);
+    }
+    if (game.currentPlayer !== user.sub) {
+      throw new HttpException(
+        'You are not allowed to restart this game. Please wait for opponent',
+        400,
+      );
     }
     this.gameService.restartGame(game);
     return this.gameService.findOne(id);
   }
 
   @Post(':id/quit')
-  quitGame(
-    @Param('id') id: string,
-    @Body() body: { playerPrivateKey: string; playerPublicKey: string },
-  ) {
+  quitGame(@Param('id') id: string, @CognitoUser() user: any) {
     const game = this.gameService.findOne(id);
     if (!game) {
       throw new HttpException('Game not found', 404);
     }
-    if (
-      !this.gameService.authPlayer(
-        game.id,
-        body.playerPrivateKey,
-        body.playerPublicKey,
-      )
-    ) {
-      throw new HttpException('Auth fail', 400);
+    if (!user.sub) {
+      throw new HttpException('User not found', 404);
     }
-    this.gameService.removePlayerFromAllGames(body.playerPrivateKey);
+    this.gameService.removePlayerFromAllGames(user.sub);
     this.gameService.removeEmptyGames();
     return { message: 'Game quit' };
   }
